feat: split homepage and marketplace into separate pages
Homepage (/) is now a clean landing page:
- Hero with "Google for OpenClaw" positioning
- Search bar that navigates to /marketplace?q=...
- 3-step "How It Works" flow
- Install CTA + footer

Marketplace (/marketplace) gets all the browsing content:
- Search with URL param support (?q=)
- Featured and trending skill sections
- Full skill grid with type/price/badge filters
- Creator CTA
