+++ server/web/src/App.jsx
@@ -1,5 +1,6 @@
 import { Routes, Route } from 'react-router-dom';
 import Skills from './pages/Skills';
+import Marketplace from './pages/Marketplace';
 import SkillDetail from './pages/SkillDetail';
 import Docs from './pages/Docs';
 import Layout from './components/Layout';
@@ -9,6 +10,7 @@ function App() {
     <Routes>
       <Route element={<Layout />}>
         <Route path="/" element={<Skills />} />
+        <Route path="/marketplace" element={<Marketplace />} />
         <Route path="/skill/:id" element={<SkillDetail />} />
         <Route path="/docs" element={<Docs />} />
       </Route>
