+++ server/web/src/index.css
@@ -3462,6 +3462,25 @@ a {
 
 /* ── MARKETPLACE ──────────────────────────────────────────────────────────── */
 
+.ub-marketplace-hero {
+  background: var(--bg-surface);
+  border-bottom: 1px solid var(--border-default);
+  padding: 100px 24px 48px;
+}
+
+.ub-marketplace-hero-inner {
+  max-width: var(--max-width);
+  margin: 0 auto;
+}
+
+.ub-marketplace-heading {
+  font-family: var(--font-display);
+  font-size: 1.5rem;
+  font-weight: 700;
+  letter-spacing: 0.1em;
+  margin-bottom: 24px;
+}
+
 .ub-marketplace {
   max-width: var(--max-width);
   margin: 0 auto;
